backend/requirements.txt
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
boto3==1.22.1
botocore==1.25.1
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
chardet==4.0.0
charset-normalizer==2.0.12
clx-sdk-xms==1.0.0
colorama==0.4.4
configparser==5.2.0
constantly==15.1.0
crayons==0.4.0
crossplane==0.5.7
cryptography==37.0.0
daphne==3.0.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.4
django-background-tasks==1.2.5
django-ckeditor==6.4.0
django-cleanup==6.0.0
django-colorfield==0.6.3
django-compat==1.0.15
django-dbbackup==3.3.0
django-filter==21.1
django-hCaptcha==0.2.0
django-import-export==2.8.0
django-js-asset==2.0.0
django-makemessages-xgettext==0.1.1
django-qr-code==3.0.0
django-ranged-response==0.2.0
django-rest-auth==0.9.5
django-rest-passwordreset==1.2.1
django-simple-captcha==0.5.17
django-storages==1.12.3
django-transadmin==0.4
django-translations==1.3.0
django-user-agents==0.4.0
django4-background-tasks==1.2.7
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
et-xmlfile==1.1.0
google-api-core==2.7.2
google-api-python-client==2.46.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.0
h11==0.13.0
httplib2==0.20.4
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
iso8601==1.0.2
jmespath==1.0.0
MarkupPy==1.14
numpy==1.22.3
oauth2client==4.1.3
odfpy==1.4.1
opencv-python==4.5.5.64
openpyxl==3.0.9
outcome==1.1.0
Pillow==9.1.0
protobuf==3.20.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==22.0.0
pyparsing==3.0.8
PySocks==1.7.1
python-dateutil==2.8.2
pytlv==0.71
pytz==2022.1
PyYAML==6.0
requests==2.27.1
rsa==4.8
s3transfer==0.5.2
segno==1.4.1
selenium==4.1.3
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tablib==3.2.1
tlv8==0.10.0
trio==0.20.0
trio-websocket==0.9.2
twilio==6.63.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.2.0
tzdata==2022.1
ua-parser==0.10.0
uritemplate==4.1.1
urllib3==1.26.9
user-agents==2.2.0
uttlv==0.6.0
webdriver-manager==3.5.4
websocket-client==1.3.2
wsproto==1.1.0
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.4.0
